requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning frameworks
torch>=1.9.0
torchvision>=0.10.0

# Medical imaging
SimpleITK>=2.1.0
nibabel>=3.2.0
opencv-python>=4.5.0

# Data processing and analysis
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.62.0

# Optional dependencies
# tensorflow>=2.6.0  # Uncomment if using TensorFlow
# monai>=1.0.0      # Uncomment if using MONAI for medical imaging
# albumentations>=1.1.0  # Uncomment if using data augmentation 